Update README.md
Added:

The **Games of the XVI Olympiad** took place in a Southern Hemisphere country for the first time from November 22 to December 8, 1956. Three thousand, three hundred and 76 athletes (2,938 men and 376 women) from 72 countries participated.

The equestrian trials were held six months earlier in Stockholm, Sweden, between June 11 and 17 due to Australian quarantine laws. Thus, the 1956 Melbourne Olympic Games were the first (and to date only) with events on different continents. 

The Melbourne Olympic Games were marked by boycotts months before the opening because of political conflicts such as the [Suez war](https://history.state.gov/milestones/1953-1960/suez) and the repression of the Hungarian revolution. However, at the closing ceremony, around 500 athletes marched together in a single group under the Olympic flag. For this reason, they were nicknamed the *friendship games*.

The **Soviet Union** was the country that won the most medals, a total of 98, and for the first time surpassed the **United States**, which won 74 medals. The locals were in third place in the medal table. **Australia** tripled the number of medals won in 1952 and achieved their best performance until Atlanta 1996.